requirements.txt
streamlit
gdown
opencv-python-headless
Pillow
numpy
tensorflow
pandas
keras
streamlit-option-menu
